Set up Cypress testing
Eventually I want to try migrating GMRA to React so we can more easily manage application state and all the different button components. Automated testing should help take some of the pain out of this. I chose the cypress framework because it seemed like the easiest to set up for testing extensions.